Use pthread mutex and mark several fields const
Using memcpy to write to a destination struct holding const fields
is questionable but done to get around there being no (JAVA-like)
final attribute to only allow a variable to be changed once.

It would be funny if one day clang/gcc refuse to compile this
because they can recognize const is being abused.